requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
psycopg2-binary
pydantic
email-validator
passlib
python-jose[cryptography]
fastapi-mail